게시글 정렬 기능 수정, 게시글 조회 개수 구현, 마이페이지 내 정보, 내가 쓴 글 보기, 내 댓글 보기 구현

+++ community/src/main/java/kr/co/community/controller/CommentController.java
@@ -48,8 +48,14 @@ public class CommentController {
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable("id") Long id) throws Exception {
+    public String delete(@PathVariable("id") Long id, Principal principal) throws Exception {
         Comment comment = commentService.findById(id);
+        Member member = memberService.findByEmail(principal.getName());
+
+        if(comment.getMember() != member) {
+            throw new Exception("삭제 권한이 없습니다.");
+        }
+
         Post post = postService.findById(comment.getPost().getId());
 
         commentService.deleteComment(comment);
